refactor: move DecoratedInput and DecoratedPasswordInput to separate folder
