+++ app/assets/javascripts/components/AccountMenu/AdvancedOptions.tsx
@@ -4,7 +4,7 @@ import { observer } from 'mobx-react-lite'
 import { FunctionComponent } from 'preact'
 import { useEffect, useState } from 'preact/hooks'
 import { Checkbox } from '../Checkbox'
-import { DecoratedInput } from '../DecoratedInput'
+import { DecoratedInput } from '../DecoratedInput/DecoratedInput'
 import { Icon } from '../Icon'
 
 type Props = {
