+++ app/assets/javascripts/components/AccountMenu/CreateAccount.tsx
@@ -5,8 +5,8 @@ import { FunctionComponent } from 'preact'
 import { StateUpdater, useEffect, useRef, useState } from 'preact/hooks'
 import { AccountMenuPane } from '.'
 import { Button } from '../Button'
-import { DecoratedInput } from '../DecoratedInput'
-import { DecoratedPasswordInput } from '../DecoratedPasswordInput'
+import { DecoratedInput } from '../DecoratedInput/DecoratedInput'
+import { DecoratedPasswordInput } from '../DecoratedInput/DecoratedPasswordInput'
 import { Icon } from '../Icon'
 import { IconButton } from '../IconButton'
 import { AdvancedOptions } from './AdvancedOptions'
