+++ app/assets/javascripts/components/AccountMenu/SignIn.tsx
@@ -7,8 +7,8 @@ import { useCallback, useEffect, useRef, useState } from 'preact/hooks'
 import { AccountMenuPane } from '.'
 import { Button } from '../Button'
 import { Checkbox } from '../Checkbox'
-import { DecoratedInput } from '../DecoratedInput'
-import { DecoratedPasswordInput } from '../DecoratedPasswordInput'
+import { DecoratedInput } from '../DecoratedInput/DecoratedInput'
+import { DecoratedPasswordInput } from '../DecoratedInput/DecoratedPasswordInput'
 import { Icon } from '../Icon'
 import { IconButton } from '../IconButton'
 import { AdvancedOptions } from './AdvancedOptions'
